Medium: sbd-md: Restructure inquisitor bootstrapping and other cleanups.
Instead of the inquisitor forking all servants once in a "trial" mode to
see if enough respond, then killing those, deciding whether to proceed
or if to exit, and then starting the "real" servants, the logic is now
as follows:

The sbd-md command forks the daemon-to-be, which immediately forks the
real servant processes. However, the frontend process lingers until it
either receives the OK from the daemon (SIG_LIVENESS after the daemon
has exceeded the quorum threshold once); or notices that the daemon
exits.

This gets rid of a few levels of nesting and special code in the servant
process.

The servant now also uses open_device(). (not sure why it didn't
before.)

Split the foreach_servants() function into one which starts and one
which kills the servants.

As part of the restructured bootstrap code, modify
check_timeout_inconsistent() to compare timeouts across all devices in
one go, skipping devices that are currently not accessible.

Limit the number of automatic servant restarts to 10; after that,
SIGUSR1 has to be sent. (I managed to bog down my test environment once
and decided this was a good idea ;-)

Minor coding style adjustments in the course of the former.  (Sorry,
they ought to be in a separate patch.) Unify list handling, some
indentation and line breaks.