pacemaker: reconnect to the CIB if an update fails to apply
This can occasionally happen. While this is a bug in pacemaker, it makes
sense for sbd to recover better from this error.

If sbd fails to reconnect within the time, the master process will still
time out the pacemaker vote, so not explicitly setting the state to
"unhealthy" here is acceptable. (And better than self-fencing due to a
really transient error.)

Also, some code refactoring.